+++ packages/shared/common/src/internal/index.ts
@@ -5,3 +5,4 @@ export * from './events';
 export * from './fdv2';
 export * from './metadata';
 export * from './plugins';
+export * from './json';