+++ packages/tooling/jest/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [0.1.7](https://github.com/launchdarkly/js-core/compare/jest-v0.1.6...jest-v0.1.7) (2025-05-22)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/react-native-client-sdk bumped from ~10.10.1 to ~10.10.2
+
+## [0.1.6](https://github.com/launchdarkly/js-core/compare/jest-v0.1.5...jest-v0.1.6) (2025-05-21)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/react-native-client-sdk bumped from ~10.10.0 to ~10.10.1
+
 ## [0.1.5](https://github.com/launchdarkly/js-core/compare/jest-v0.1.4...jest-v0.1.5) (2025-04-29)